Working flask app
