+++ templates/index.html
@@ -22,10 +22,6 @@
           Playlist Description:
           <input type="text" name="desc" />
         </label><br/>
-        <label>
-          API Key
-          <input type="text" name="apikey" />
-        </label><br/>
         <label>Artist Song Separator   </label>
         <select id="delimiter" name="delimiter">
           <option value="comma">comma (,)</option>
